+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [ Unreleased ] - ReleaseDate
+
+### Changed
+
+- Raised MSRV to 1.70.0 to remove `lazy_static` dependency
+  ([#550](https://github.com/asomers/mockall/pull/550))
 
 ## [ 0.12.1 ] - 2023-12-21